SingleNumber, Robot Movement and AddBinary
Single Number returns the missing number in a list of numbers where all numbers are duplicated once except for one number.

Robot movement reads a string moving a robot from (0, 0) on the cartesian plane with UDLR for Up Down Left Right actions. where each Character is an action of 1 block.
Returns a boolean if the robot is back to (0, 0) at the end of the command string.

Add binary accepts 2 binary numbers encoded as strings and returns the sum